optional suggestion implemented
